requirements.txt
feedparser==6.0.11
pandas==2.2.2
numpy==1.26.4
pydantic==2.8.2
sqlalchemy==2.0.36
python-dateutil==2.9.0
APScheduler==3.10.4
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.2.1
scikit-learn==1.5.2
matplotlib==3.9.2
streamlit==1.38.0
PyYAML==6.0.2
fastapi==0.115.0
uvicorn==0.30.6
